gui new files and hierarchy
