+++ gui/include/local.hpp
@@ -1,0 +1,26 @@
+/******************************************************************************/
+/*	Project:	Remote Tasker   										      */
+/*	File:		local.hpp	    										          */
+/*	Version: 	0.03														  */
+/******************************************************************************/
+#ifndef __LOCAL_HPP__
+#define __LOCAL_HPP__
+
+#include <iostream>
+#include <vector>
+
+#include <gtkmm.h>
+
+namespace remote_tasker
+{
+
+class Local
+{
+    Local();
+
+
+}
+
+} // namespace remote_tasker
+
+#endif // __LOCAL_HPP__
